+++ batch/__init__.py
@@ -1,0 +1,2 @@
+APP_LABEL = "batch"
+
