--- externalConnect/__init__.py
@@ -1,2 +1,0 @@
-# 外部接続用アプリケーション
-APP_LABEL = "externalConnect"
