--- externalConnect/views.py
@@ -1,1 +1,0 @@
-# Create your views here.
