+++ mymodule/external_connector/baseConnector.py
