--- mymodule/mongoRepository/tradeHistDao.py
@@ -1,5 +1,0 @@
-from mymodule.mongoRepository.mongoDaoBase import MongoDaoBase
-
-
-# class TradeHistDao(MongoDaoBase):
-
