+++ mymodule/mongo_repository/dao/mongoDaoBase.py
