+++ mymodule/mongo_repository/dao/tradeHistDao.py
@@ -1,0 +1,5 @@
+from mymodule.mongo_repository.mongoDaoBase import MongoDaoBase
+
+
+# class TradeHistDao(MongoDaoBase):
+
