+++ mymodule/mongo_repository/entity/models.py
